requirements.txt
python-dotenv
google-genai
streamlit==1.31.1
pandas
prophet==1.1.5
scikit-learn
uvicorn
fastapi[standard]
requests
